Added original grf file.
This will be overwritten, but is licensed under CC BY-NC-SA 3.0 in the same way as the original file. The source code of all GRFs is dual-licensed due to the use of a GPL v2 compiler, but the graphics which will be re-implemented are exclusively under the Creative Commons License as specified by Ufiby.